requirements.txt
# Core
openai>=1.30.0
python-telegram-bot>=21.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
tiktoken>=0.7.0

# Optional: Anthropic provider
# anthropic>=0.25.0

# Optional: Web search
# tavily-python>=0.3.0
